Support configuration using json
json is now an alternative way to configure the bots variables like TOKEN or PREFIX.
If a parameter is given using env-variable and json the one from env will be preferred.

I is suggested to continue using env variables, but it's neat to have both, especially for beginners.
The path to the file is hard-coded as ./data/config.json at the moment